prelim-3 SIR changed for over two day incubation
Ok, I was sampling after midnight, so i changed the code in calculate_sir_alex to include date... hopefully this will not interfere with input raw csvs that just have hours. right now the input format is set as MM/DD/YYYY hh/MM (hh as in 24-hour.. even putting am/pm made it so that it was 28 hours incubation time instead of 4) these are mostly lines 91 and 92. also changed for select function needs "dplyr::"